Updated to use new layout of TCK-classes
