+++ src/main/asciidoc/dataobjects.adoc
@@ -49,7 +49,8 @@ Set the maximum number of queries when an hostname is resolved.
 +++
 |[[ndots]]`ndots`|`Number (int)`|
 +++
-Set the ndots value used when resolving using search domains, the default value is <code>1</code>.
+Set the ndots value used when resolving using search domains, the default value is <code>-1</code> which
+ determines the value from the OS on Linux or uses the value <code>1</code>.
 +++
 |[[optResourceEnabled]]`optResourceEnabled`|`Boolean`|
 +++
@@ -164,6 +165,11 @@ Set the idle timeout, in seconds. zero means don't timeout.
 |[[keyStoreOptions]]`keyStoreOptions`|`link:dataobjects.html#JksOptions[JksOptions]`|
 +++
 Set the key/cert options in jks format, aka Java keystore.
++++
+|[[localAddress]]`localAddress`|`String`|
++++
+Set the local interface to bind for network connections. When the local address is null,
+ it will pick any local address, the default local address is null.
 +++
 |[[logActivity]]`logActivity`|`Boolean`|
 +++
@@ -574,7 +580,7 @@ Set the additional debug data
 +++
 |[[errorCode]]`errorCode`|`Number (long)`|
 +++
-@return the  error code
+
 +++
 |[[lastStreamId]]`lastStreamId`|`Number (int)`|
 +++
@@ -716,6 +722,11 @@ Set whether keep alive is enabled on the client
 |[[keyStoreOptions]]`keyStoreOptions`|`link:dataobjects.html#JksOptions[JksOptions]`|
 +++
 Set the key/cert options in jks format, aka Java keystore.
++++
+|[[localAddress]]`localAddress`|`String`|
++++
+Set the local interface to bind for network connections. When the local address is null,
+ it will pick any local address, the default local address is null.
 +++
 |[[logActivity]]`logActivity`|`Boolean`|
 +++
@@ -1017,7 +1028,7 @@ Set the websocket subprotocols supported by the server.
 ^|Name | Type ^| Description
 |[[alpnAvailable]]`alpnAvailable`|`Boolean`|
 +++
-@return if alpn support is available via the JDK SSL engine
+
 +++
 |===
 
@@ -1130,6 +1141,11 @@ Set the idle timeout, in seconds. zero means don't timeout.
 |[[keyStoreOptions]]`keyStoreOptions`|`link:dataobjects.html#JksOptions[JksOptions]`|
 +++
 Set the key/cert options in jks format, aka Java keystore.
++++
+|[[localAddress]]`localAddress`|`String`|
++++
+Set the local interface to bind for network connections. When the local address is null,
+ it will pick any local address, the default local address is null.
 +++
 |[[logActivity]]`logActivity`|`Boolean`|
 +++
@@ -1348,7 +1364,6 @@ Set whether Netty pooled buffers are enabled
 == NetworkOptions
 
 ++++
- @author <a href="http://tfox.org">Tim Fox</a>
 ++++
 '''
 
@@ -1446,11 +1461,11 @@ Set whether the file is to be opened for writing
 ^|Name | Type ^| Description
 |[[alpnAvailable]]`alpnAvailable`|`Boolean`|
 +++
-@return when alpn support is available via OpenSSL engine
+
 +++
 |[[available]]`available`|`Boolean`|
 +++
-@return when OpenSSL is available
+
 +++
 |[[sessionCacheEnabled]]`sessionCacheEnabled`|`Boolean`|
 +++
@@ -1516,7 +1531,7 @@ Sets whether or not this option can receive several values.
 +++
 |[[name]]`name`|`String`|
 +++
-@return the option name. It returns the long name if set, the short name otherwise. It cannot return <code>null</code> for valid option
+
 +++
 |[[required]]`required`|`Boolean`|
 +++
