+++ src/main/scala/io/vertx/scala/core/http/HttpServerResponse.scala
@@ -96,8 +96,7 @@ class HttpServerResponse(private val _asJava: io.vertx.core.http.HttpServerRespo
   }
 
   /**
-    * @return the HTTP status message of the response. If this is not specified a default value will be used depending on what
-    * [[io.vertx.scala.core.http.HttpServerResponse#setStatusCode]] has been set to.
+    * @return the HTTP status message of the response. If this is not specified a default value will be used depending on what [[#setStatusCode]] has been set to.
     */
   def getStatusMessage(): String = {
     _asJava.getStatusMessage()
@@ -375,9 +374,10 @@ class HttpServerResponse(private val _asJava: io.vertx.core.http.HttpServerRespo
   }
 
   /**
-    * Provide a handler that will be called just before the last part of the body is written to the wire
-    * and the response is ended.
-    * This provides a hook allowing you to do any more operations before this occurs.
+    * Provides a handler that will be called after the last part of the body is written to the wire.
+    * The handler is called asynchronously of when the response has been received by the client.
+    * This provides a hook allowing you to do more operations once the request has been sent over the wire
+    * such as resource cleanup.
     * @param handler the handler
     * @return a reference to this, so the API can be used fluently
     */
@@ -394,7 +394,7 @@ class HttpServerResponse(private val _asJava: io.vertx.core.http.HttpServerRespo
   }
 
   /**
-    * @return the id of the stream of this response,  for HTTP/1.x
+    * @return the id of the stream of this response, {@literal -1` for HTTP/1.x
     */
   def streamId(): Int = {
     _asJava.streamId()
