+++ src/main/scala/io/vertx/scala/core/http/WebSocketFrame.scala
@@ -57,8 +57,7 @@ class WebSocketFrame(private val _asJava: io.vertx.core.http.WebSocketFrame) {
   }
 
   /**
-    * @return the content of this frame as a UTF-8 string and returns the
-    * converted string. Only use this for text frames.
+    * @return the content of this frame as a UTF-8 string and returns the converted string. Only use this for text frames.
     */
   def textData(): String = {
     if (cached_0 == null) {
