+++ src/main/scala/io/vertx/scala/core/streams/WriteStream.scala
@@ -63,8 +63,12 @@ trait WriteStream[T]
 
   /**
   * Set the maximum size of the write queue to `maxSize`. You will still be able to write to the stream even
-  * if there is more than `maxSize` bytes in the write queue. This is used as an indicator by classes such as
+  * if there is more than `maxSize` items in the write queue. This is used as an indicator by classes such as
   * `Pump` to provide flow control.
+  * <p/>
+  * The value is defined by the implementation of the stream, e.g in bytes for a
+  * [[io.vertx.scala.core.net.NetSocket]], the number of [[io.vertx.scala.core.eventbus.Message]] for a
+  * [[io.vertx.scala.core.eventbus.MessageProducer]], etc...
   * @param maxSize the max size of the write stream
   * @return a reference to this, so the API can be used fluently
   */
@@ -78,7 +82,10 @@ trait WriteStream[T]
 
   /**
   * Set a drain handler on the stream. If the write queue is full, then the handler will be called when the write
-  * queue has been reduced to maxSize / 2. See [[io.vertx.scala.core.streams.Pump]] for an example of this being used.
+  * queue is ready to accept buffers again. See [[io.vertx.scala.core.streams.Pump]] for an example of this being used.
+  * <p/>
+  * The stream implementation defines when the drain handler, for example it could be when the queue size has been
+  * reduced to `maxSize / 2`.
   * @param handler the handler
   * @return a reference to this, so the API can be used fluently
   */
@@ -135,8 +142,12 @@ object WriteStream {
 
     /**
       * Set the maximum size of the write queue to `maxSize`. You will still be able to write to the stream even
-      * if there is more than `maxSize` bytes in the write queue. This is used as an indicator by classes such as
+      * if there is more than `maxSize` items in the write queue. This is used as an indicator by classes such as
       * `Pump` to provide flow control.
+      * <p/>
+      * The value is defined by the implementation of the stream, e.g in bytes for a
+      * [[io.vertx.scala.core.net.NetSocket]], the number of [[io.vertx.scala.core.eventbus.Message]] for a
+      * [[io.vertx.scala.core.eventbus.MessageProducer]], etc...
       * @param maxSize the max size of the write stream
       * @return a reference to this, so the API can be used fluently
       */
@@ -155,7 +166,10 @@ object WriteStream {
 
     /**
       * Set a drain handler on the stream. If the write queue is full, then the handler will be called when the write
-      * queue has been reduced to maxSize / 2. See [[io.vertx.scala.core.streams.Pump]] for an example of this being used.
+      * queue is ready to accept buffers again. See [[io.vertx.scala.core.streams.Pump]] for an example of this being used.
+      * <p/>
+      * The stream implementation defines when the drain handler, for example it could be when the queue size has been
+      * reduced to `maxSize / 2`.
       * @param handler the handler
       * @return a reference to this, so the API can be used fluently
       */
